+++ Samples/Client/Android-Native/3dToolkitAndroid/app/src/test/java/microsoft/a3dtoolkitandroid/ObserverTests.java
@@ -9,13 +9,8 @@ import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.isNotNull;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.timeout;
 import static org.mockito.Mockito.verify;
 import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.webrtc.IceCandidate;
-import org.webrtc.SessionDescription;
 
 
 public class ObserverTests {
@@ -23,11 +18,11 @@ public class ObserverTests {
     private static final String DUMMY_SDP = "sdp";
     public static final int SERVER_WAIT = 100;
     public static final int NETWORK_TIMEOUT = 1000;
-    ServerListActivity.PeerConnectionObserver peerConnectionObserver;
-    ServerListActivity.SDPObserver sdpObserver;
+    ConnectActivity.PeerConnectionObserver peerConnectionObserver;
+    ConnectActivity.SDPObserver sdpObserver;
     @Before
     public void setUp() {
-        peerConnectionObserver = mock(ServerListActivity.PeerConnectionObserver.class);
-        sdpObserver = mock(ServerListActivity.SDPObserver.class);
+        peerConnectionObserver = mock(ConnectActivity.PeerConnectionObserver.class);
+        sdpObserver = mock(ConnectActivity.SDPObserver.class);
     }
 }
